requirements.txt
asgiref==3.5.2
autopep8==2.0.0
Django==4.1.2
djangorestframework==3.14.0
factory-boy==3.2.1
Faker==15.1.3
protobuf==3.19.4
pycodestyle==2.9.1
python-dateutil==2.8.2
pytz==2022.5
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
